merage

+++ src/main/java/com/avalok/ib/handler/HistoricalDataHandler.java
@@ -10,17 +10,18 @@ import com.alibaba.fastjson.JSONObject;
 import com.bitex.util.Redis;
 
 public class HistoricalDataHandler implements ApiController.IHistoricalDataHandler {
-	protected final JSONArray historyBar = new JSONArray();
-	private Long id;
-	
-	public HistoricalDataHandler(Long _id) {
-		id = _id;
-	}
-	
+    protected final JSONArray historyBar = new JSONArray();
+    private Long id;
+
+    public HistoricalDataHandler(Long _id) {
+        id = _id;
+    }
+
     @Override
     public void historicalData(Bar bar) {
 //      String msg = EWrapperMsgGenerator.historicalData(0,String.valueOf(bar.time()), bar.open(), bar.high(), bar.low(), bar.close(), bar.volume(), bar.count(), bar.wap());
 //    	String msg = bar.toString();
+        // log("len: "+ historyBar.size());
 
         JSONObject j = new JSONObject();
         j.put("timestamp", bar.time());
@@ -28,7 +29,7 @@ public class HistoricalDataHandler implements ApiController.IHistoricalDataHandl
         j.put("high", bar.high());
         j.put("low", bar.low());
         j.put("close", bar.close());
-        j.put("volume", bar.volume());
+        j.put("volume", bar.volume().longValue());
 //        j.put("formattedTime", bar.formattedTime());
         historyBar.add(j);
 //    	log(bar.formattedTime() + " " + msg);
@@ -37,9 +38,9 @@ public class HistoricalDataHandler implements ApiController.IHistoricalDataHandl
     @Override
     public void historicalDataEnd() {
 //    	Redis.setex(null, 30, historyBar);
-		String key = "IBGateway:ReqIdHistoricalData:" + id;
-		log("Redis -> " + key);
-		Redis.setex(key, 30, historyBar);
+        String key = "IBGateway:ReqIdHistoricalData:" + id;
+        log("Redis -> " + key);
+        Redis.setex(key, 30, historyBar);
 
     }
 }
